Require Java 11 (minimum) at run time
