+++ version/src/main/java/io/smallrye/common/version/JDKSpecific.java
@@ -1,14 +1,16 @@
 package io.smallrye.common.version;
 
+import java.lang.module.ModuleDescriptor;
+
 final class JDKSpecific {
     private JDKSpecific() {
     }
 
     static boolean hasJpms() {
-        return false;
+        return true;
     }
 
     static int compareJpms(String v1, String v2) {
-        throw new UnsupportedOperationException("Java 9 only");
+        return Integer.signum(ModuleDescriptor.Version.parse(v1).compareTo(ModuleDescriptor.Version.parse(v2)));
     }
 }
